requirements.txt
beautifulsoup4==4.13.3
python-docx
Flask==3.1.0
Flask_Login==0.6.3
flask_sqlalchemy==3.1.1
joblib==1.4.2
nltk==3.9.1
pandas==2.2.3
PyPDF2==3.0.1
requests==2.32.3
scikit-learn==1.4.2
gunicorn
lxml
setuptools